django_start/requirements.dev.txt
-r requirements.txt

pylint
pylint-django

pytest
pytest-django
pytest-asyncio

coverage
pytest-cov
pytest-aiohttp
pytest-mock
pytest-env